[app] converted attach roles modal to view
