+++ packages/app/src/components/App.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom"
 import RolesView from "./views/RolesList/RolesView"
-import AttachRolesModifierModal from "./modals/AttachRolesModifierModal"
+import AttachRolesModifierView from "./views/AttachRolesModifier/AttachRolesModifierView"
 import RoleView from "./views/Role/RoleView"
 import { Root } from "./Root"
 
@@ -10,7 +10,7 @@ export const App = (): React.ReactElement => {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Root />}>
-          <Route index element={<AttachRolesModifierModal onClose={() => {}} />} />
+          <Route index element={<AttachRolesModifierView />} />
           <Route path=":module" element={<RolesView />} />
           <Route path=":module/roles/:roleId" element={<RoleView />} />
           <Route path="*" element={<Navigate to="/" replace />} />
